% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoV.R
\name{CoV}
\alias{CoV}
\title{CoV}
\usage{
CoV(x, na.rm = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{Numeric Vector}

\item{na.rm}{TRUE or FALSE}

\item{verbose}{TRUE or FALSE}
}
\value{
A vector containing the coefficient of variance
}
\description{ The 'CoV' package calculates the coefficient of variance of a specific variable/column. It calculates the CoV by first calculating the standard deviation and mean using sd() and mean(), then multiplying the value by 100.
}
\examples{
CoV(1:5)
CoV(-100:2000)
}
